timeseries.io.gwf: simplified multi-file GWF read
use the `read_multi` infrastructure used for `EventTable.read`
